+++ src/main/java/org/prebid/server/bidder/BidderRequester.java
@@ -5,6 +5,8 @@ import io.vertx.core.Future;
 import org.prebid.server.bidder.model.BidderSeatBid;
 import org.prebid.server.execution.Timeout;
 
+import java.math.BigDecimal;
+
 /**
  * Interface for sending requests to Bidders.
  */
@@ -13,5 +15,5 @@ public interface BidderRequester {
     /**
      * Executes given request to a bidder.
      */
-    Future<BidderSeatBid> requestBids(BidRequest bidRequest, Timeout timeout, Float bidPriceAdjustmentFactor);
+    Future<BidderSeatBid> requestBids(BidRequest bidRequest, Timeout timeout, BigDecimal bidPriceAdjustmentFactor);
 }
